backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-SocketIO==5.3.6
pymongo==4.5.0
bcrypt==4.0.1
python-dotenv==1.0.0
pandas==2.1.1
openpyxl==3.1.2
matplotlib==3.7.2
seaborn==0.12.2
numpy==1.24.3
Pillow==10.0.1
eventlet==0.33.3
psutil==7.0.0
gunicorn==21.2.0 